+++ src/main/java/jj2000/j2k/quantization/dequantizer/StdDequantizer.java
@@ -349,8 +349,8 @@ public class StdDequantizer extends Dequantizer {
                 // code-block width.
                 for (j=outiarr.length-1; j>=0; j--) {
                     temp = outiarr[j]; // input array is same as output one
-                    outiarr[j] = (int)(((float)((temp >= 0) ? temp :
-                                                -(temp&0x7FFFFFFF)))*step);
+                    outiarr[j] = (int)(((temp >= 0) ? temp :
+                                                -(temp&0x7FFFFFFF))*step);
                 }
                 break;
             case DataBlk.TYPE_FLOAT:
@@ -362,8 +362,8 @@ public class StdDequantizer extends Dequantizer {
                          jmin = w*(h-1); j>=0; jmin -= w) {
                     for (; j>=jmin; k--, j--) {
                         temp = inarr[k];
-                        outfarr[j] = ((float)((temp >= 0) ? temp :
-                                              -(temp&0x7FFFFFFF)))*step;
+                        outfarr[j] = ((temp >= 0) ? temp :
+                                              -(temp&0x7FFFFFFF))*step;
                     }
                     // Jump to beggining of previous line in input
                     k -= inblk.scanw - w;
